Split out the mega-index.ts
