+++ commands/item.ts
@@ -1,0 +1,27 @@
+import { SlashCommandBuilder, SlashCommandIntegerOption, SlashCommandStringOption } from "discord.js";
+
+export const item_command = new SlashCommandBuilder()
+    .setName('item')
+    .setDescription('Create an item thread')
+    .addIntegerOption(new SlashCommandIntegerOption()
+      .setName('number')
+      .setDescription('Item Number')
+      .setRequired(true)
+    )
+    .addIntegerOption(new SlashCommandIntegerOption()
+      .setName('page')
+      .setDescription('Page Number')
+      .setRequired(true)
+    )
+    .addStringOption(new SlashCommandStringOption()
+      .setRequired(false)
+      .setName('category')
+      .setDescription('Which list category does this belong to (if any)')
+      .setChoices({name: "Foo", value: "bar"})
+    )
+    .addStringOption(new SlashCommandStringOption()
+      .setName('summary')
+      .setDescription('Item summary to be included in channel name')
+      .setMaxLength(80)
+      .setRequired(false)
+    );
