+++ package.json
@@ -13,8 +13,5 @@
     "sequelize": "^6.37.7",
     "sqlite3": "^5.1.7",
     "typescript": "^5.8.3"
-  },
-  "devDependencies": {
-    "sequelize-cli": "^6.6.2"
   }
 }
